Server Api's Refactor and Updated with Proper Response Status

+++ Server/Controllers/AdminController/ManageHCPController.cs
@@ -4,7 +4,6 @@ using imc_web_api.Dtos.AdminDtos.HCPDtos;
 //using imc_web_api.Dtos.HCPDtos;
 using imc_web_api.Models;
 using imc_web_api.Service.AdminServices.ManageHCPServices;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace imc_web_api.Controllers.AdminController
@@ -28,36 +27,67 @@ namespace imc_web_api.Controllers.AdminController
         //-->Add HCP
         [HttpPost]
         [Route("AddHCP")]
-        [Authorize(Roles = "Admin,ServiceProvider")]
         public async Task<IActionResult> AddHCP([FromBody] HCPRequestDTO UserInputReguest)
         {
-            var HCP_Model = _mapper.Map<serviceprovidertype>(UserInputReguest);
-            var HCP_Result = await _manageHCPService.AddProvider(HCP_Model);
-
-            var HCPDto_Result = _mapper.Map<HCPRequestDTO>(HCP_Result);
-
-            return Ok(new
+            try
             {
-                success = true,
-                Message = "Provider Added Successfully!",
-                Data = HCPDto_Result,
-            });
+                if (UserInputReguest == null)
+                {
+                    return BadRequest(ModelState);
+                }
+                var HCP_Model = _mapper.Map<serviceprovidertype>(UserInputReguest);
+                var HCP_Result = await _manageHCPService.AddProvider(HCP_Model);
+
+                var HCPDto_Result = _mapper.Map<HCPRequestDTO>(HCP_Result);
+
+                return Ok(new
+                {
+                    success = true,
+                    Message = "Provider Added Successfully!",
+                });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new
+                {
+                    Success = false,
+                    ErrorMessage = ex.Message
+                });
+            }
         }
 
         //-->Update HCP
         [HttpPut]
         [Route("UpdateHCP/{id:Guid}")]
-        [Authorize(Roles = "Admin,ServiceProvider")]
         public async Task<IActionResult> UpdateHCP(Guid id, [FromBody] HCPRequestDTO InputRequest)
         {
-            var HCP_Model_Result = await _manageHCPService.UpdateProvider(id, InputRequest);
-            var HCP_DTO_Result = _mapper.Map<HCPResponseDTO>(HCP_Model_Result);
-            return Ok(new
+            try
             {
-                success = true,
-                Data = HCP_DTO_Result,
-                Message = "Uppdated Record Successfully!"
-            });
+                if (InputRequest == null || id == Guid.Empty)
+                {
+                    return BadRequest("Input Field is Required ");
+                }
+                var HCP_Model_Result = await _manageHCPService.UpdateProvider(id, InputRequest);
+
+                if (HCP_Model_Result == null)
+                {
+                    return NotFound("Record Not Found!");
+                }
+                var HCP_DTO_Result = _mapper.Map<HCPResponseDTO>(HCP_Model_Result);
+                return Ok(new
+                {
+                    success = true,
+                    Message = "Health care Record Updated!"
+                });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new
+                {
+                    Success = false,
+                    ErrorMessage = ex.Message
+                });
+            }
         }
 
         //-->GetAll HCP
@@ -65,14 +95,25 @@ namespace imc_web_api.Controllers.AdminController
         [Route("GetHCPs")]
         public async Task<IActionResult> GetHCPs()
         {
-            var HCP_Model_Result = await _manageHCPService.GetProviders();
-            var HCP_DTO_Result = _mapper.Map<List<HCPResponseDTO>>(HCP_Model_Result);
-
-            return Ok(new
+            try
             {
-                success = true,
-                Data = HCP_DTO_Result,
-            });
+                var HCP_Model_Result = await _manageHCPService.GetProviders();
+                var HCP_DTO_Result = _mapper.Map<List<HCPResponseDTO>>(HCP_Model_Result);
+
+                return Ok(new
+                {
+                    Success = true,
+                    Data = HCP_DTO_Result,
+                });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new
+                {
+                    Success = false,
+                    ErrorMessage = ex.Message
+                });
+            }
         }
 
         // -->Get HCP By Id
@@ -80,29 +121,66 @@ namespace imc_web_api.Controllers.AdminController
         [Route("GetHCP/{id:Guid}")]
         public async Task<IActionResult> GetHCPById(Guid id)
         {
-            var HCP_Model_Result = await _manageHCPService.GetProviderById(id);
-            var HCP_DTO_Result = _mapper.Map<HCPResponseDTO>(HCP_Model_Result);
-            return Ok(new
+            try
             {
-                success = true,
-                Data = HCP_DTO_Result,
-            });
+                if (id == Guid.Empty)
+                {
+                    return BadRequest("Id Field is Required!s");
+                }
+                var HCP_Model_Result = await _manageHCPService.GetProviderById(id);
+
+                var HCP_DTO_Result = _mapper.Map<HCPResponseDTO>(HCP_Model_Result);
+
+                return Ok(new
+                {
+                    success = true,
+                    Data = HCP_DTO_Result,
+                });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new
+                {
+                    Success = false,
+                    ErrorMessage = ex.Message
+                });
+            }
         }
 
         //-->Delete HCP
         [HttpDelete]
         [Route("DeleteHCP")]
-        [Authorize(Roles = "Admin,ServiceProvider")]
         public async Task<IActionResult> DeleteHCP(Guid id)
         {
-            var HCP_Model_Result = await _manageHCPService.DeleteProvider(id);
-            var HCP_DTO_Result = _mapper.Map<HCPResponseDTO>(HCP_Model_Result);
-            return Ok(new
+            try
             {
-                success = true,
-                Data = HCP_DTO_Result,
-                Message = " Record Deleted Successfully!"
-            });
+                if (id == Guid.Empty)
+                {
+                    return BadRequest("Id Field is Required!s");
+                }
+
+                var HCP_Model_Result = await _manageHCPService.DeleteProvider(id);
+
+                if (HCP_Model_Result == null)
+                {
+                    return NotFound("Record Not Found!");
+                }
+                var HCP_DTO_Result = _mapper.Map<HCPResponseDTO>(HCP_Model_Result);
+
+                return Ok(new
+                {
+                    Success = true,
+                    Message = "Health Care Provider Deleted Successfully!"
+                });
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new
+                {
+                    Success = false,
+                    ErrorMessage = ex.Message
+                });
+            }
         }
     }
 }
